refactor(verify): register Helm settings flags
Helm's cli.EnvSettings type is responsible for registering and parsing configurations; chart-verifier uses it to parse the same high level options to offer the same semantics as Helm.

This change is required to propagate KubeConfig, KubeContext and others when installing the chart and performing chart tests in the near future.