refactor(charttesting): copy private methods from the chart-testing project
The chart-testing project orchestrates installations, upgrades and
tests of Helm charts, which is quite helpful but some useful methods
of the Testing type are private, such as doUpgrade(), doInstall() and
others so this change brings those methods as functions to provide the
required orchestration for chart-verifier.

This change brings the equivalent of 'testing.InstallChart' without
support for upgrade tests as of yet.