refactor(helm): add chart name to resulting ChartCacheItem's Path
Using Helm's 'SaveDir' function will unpack the chart into a directory
within the chartCacheDir, so the resulting 'Path' should be the
concatenation of the cache dir and the name.

This change is required to properly open a chart when it was cached
from a file or directory in the filesystem.