Set openshiftVersion value through a CLI flag (#115)
* Set openshiftVersion value through a CLI flag

The `oc version` command gives the `openshiftVersion` value if the
logged-in user (role) has access to get values of `clusteroperators`
(a cluster scoped resource in the `config.openshift.io` API group).
A developer may not have these kinds of access to run the chart-verifier.

This solution introduces a command-line flag to set the value of
`openshiftVersion`.  The flag is `--openshift-version` (`-V` as a short
option).  If the user has access to get values of `clusteroperators`
resource, that value will get precedence over the CLI flag.

Fixes #111