refactor: capture output of a  process that triggered an error
The default 'chart-testing' implementation of
'RunProcessInDirAndCaptureOutput' doesn't return the captured output
in the case the process returns an error. This isn't very useful for
chart-verifier's use case since the ignored information is important
to its users.

This change builds on the chart-verifier's version of 'tool.Helm' and
overrides methods defined by 'exec.ProcessExecutor'.