cmd/check: add check command to parse report.yaml
As part of work to add profile support in chart-verifier, we need to
move the hardcoded logic from chart submission workflow to chart-verifier.

An example is the
https://github.com/openshift-helm-charts/charts/blob/main/scripts/src/chartprreview/verify-report.sh#L197
where we hardcoded mandatory checks. Here we will change to use profiles
defined within verifier to minimize disparity. The next step would be to move the `getAnnotations` logic to verifier.

Signed-off-by: Allen Bai <abai@redhat.com>