Print step's output paramaters also to sdtout
This commit makes sure that all steps' output parameters (i.e.
GITHUB_OUTPUT) are also printed to stdout, making it easier to debug
workflow runs.

It also modifies the scripts output logic to only write to GITHUB_OUTPUT
if the environment variable is defined, making it easier to run the
scripts locally.

close #388

Signed-off-by: mgoerens <mgoerens@redhat.com>