Expire commit container images after 1 week
Prior to podman 4.6.0, there is a bug that would cause the expiration
tag to be added to the "main" container image as well in the main.yaml
workflow. See the buildah bug report:
https://github.com/containers/buildah/issues/4632

As a workaround, the --no-cache flag is used to build container images
with the "build-image" make target.

This workaround, can be removed when we use a runner image that has
podman>=4.6.0. Currently, the ubuntu-latest runner image is currently
based on Ubuntu 22.04 and had podman 3.4.4.

close #397

Signed-off-by: mgoerens <mgoerens@redhat.com>